Refactored the iTip broker so we can more easily add VTODO support.
